[fix] Copy gateway logs on interrupt (#754)
On keyboard interrupt for a transfer, gateway logs are not copied so
diagnosing status of transfers can be difficult. This is part one of a
deprovisioning error fix for keyboard interrupts.
